Accept version module names
We have a bunch of versioned REST APIs, and we usually name the modules
with that version, like `V1` and `V2` to represent the API version.
Ebert always comments in our pull requests saying that it may be a bad
module name.
We agreed that it is ok to name our modules with a `V` followed by a
number, so I've added a regex in our reek configuration file to allow
this.